# Remove arguments of _forward_eval_ϵ