They are always the same anyway. I wonder if there is any performance benefit of saving these interpret calls. Is there a standard benchmark used for checking for regressions ?
Maybe https://github.com/lanl-ansi/rosetta-opf/blob/main/jump.jl ?