@odow 2025-04-25
Yes that, with a much larger instance, say https://github.com/power-grid-lib/pglib-opf/blob/master/pglib_opf_case30000_goc.m