@odow 2025-04-25
Or build that model and then manually evaluate the Hessian with `MOI.eval_hessian_lagrangian` etc.