@odow 2025-04-25
When I refactored this out of JuMP, I tried to make the fewest possible changes. The code can be cleaned up lots.